backend/requirements.txt
deta[async]==1.1.0a2
fastapi
uvicorn
pyjwt
passlib[bcrypt]
transliterate
Pillow
python-multipart
python-dotenv
email-validator